Implement GPU occlusion culling
Add occlusion culling on the GPU to the material system. This works by creating a depth-reduction image from the depth buffer, then using it on the next frame. The bounding sphere of each surface is projected into a screen-space AABB, then its depth is tested against the depth of 4 texels in the depth image at appropriate mip-level, determined by the size of the AABB on screen. This doesn't include portal views because they use stencil buffer, so doing depth reduction and checking against depth image there would be complicated and likely require some sort of visibility buffer.